requirements.txt
python-dotenv==0.19.2
requests==2.28.0